requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
gunicorn==23.0.0
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
redis==5.3.1
taskiq==0.12.2
taskiq-redis==1.0.7
qdrant-client==1.12.1
sentence-transformers==3.3.1
anthropic==0.42.0
openai==1.59.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
httpx==0.28.1
playwright==1.49.1
pydantic-settings==2.7.1
loguru==0.7.3
sentry-sdk[fastapi]==2.19.2
pytest==8.3.4
pytest-asyncio==0.25.2
aiosqlite==0.20.0
email-validator==2.2.0
